pyproject.toml
[tool.poetry]
name = "base-tgbot"
version = "0.1.0"
description = ""
authors = ["LostUnion <nsk.19980513@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.15.0"
python-decouple = "^3.8"
hiredis = "^3.1.0"
flake8 = "^7.1.1"
loguru = "^0.7.3"
arq = "^0.26.1"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
